Security: Fix XXE/ReDoS CVEs, harden CORS config, upgrade Spring Boot to 3.4.13
- Upgrade spring-boot-starter-parent to 3.4.13 (resolves high CVEs)
- Fix ReDoS vulnerability in JwtAuthenticationFilter & JwtTokenProvider
- Harden SecurityConfig with proper CORS allowed origins (zplusejobs.com)
- Add CORS preflight handling in WebMvcConfig and nginx.prod.conf
- Fix GlobalExceptionHandler and RestExceptionHandler for safe error responses
- Update AuthRestController for CORS-safe auth endpoints
- Fix JobPostActivityRepository native query for Linux case-sensitivity
- Update frontend api.js base URL to production domain
- Update nginx.conf and nginx/nginx.prod.conf with zplusejobs.com domain

+++ backend/src/main/java/com/jobportal/config/JwtTokenProvider.java
@@ -24,10 +24,10 @@ public class JwtTokenProvider {
     private String jwtSecret;
 
     @Value("${app.jwt.expiration}")
-    private int jwtExpirationInMs;
+    private long jwtExpirationInMs;
 
     @Value("${app.jwt.refresh-expiration}")
-    private int refreshExpirationInMs;
+    private long refreshExpirationInMs;
 
     private SecretKey getSigningKey() {
         byte[] keyBytes = Decoders.BASE64.decode(jwtSecret);
